Showed the primary speaker in group control and added per-row loading spinners
